Shift Binary Heap tests to Jest
